+++ cache/ft_reset.go
@@ -31,7 +31,7 @@ Example usage:
 		log.Fatalf("Error resetting FullText cache: %v", err)
 	}
 */
-func (c *Cache) ResetFT(maxWords int, maxSizeBytes int, schema map[string]bool) error {
+func (c *Cache) ResetFT(maxWords int, maxSizeBytes uintptr, schema map[string]bool) error {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 	return c.resetFT(maxWords, maxSizeBytes, schema)
@@ -50,7 +50,7 @@ If the FullText cache is not already initialized, an error will be returned.
 @Returns:
   - error: an error if the FullText cache is not initialized or there is an issue initializing the cache with the specified parameters
 */
-func (c *Cache) resetFT(maxWords int, maxSizeBytes int, schema map[string]bool) error {
+func (c *Cache) resetFT(maxWords int, maxSizeBytes uintptr, schema map[string]bool) error {
 	// If the FT cache is not initialized, return an error
 	if c.ft == nil {
 		return fmt.Errorf("full text is not initialized")
@@ -75,7 +75,7 @@ If an error occurs during the process, it will return an error.
 @Returns:
   - error: An error if the FullText cache is not initialized or if there is an error during the reset process
 */
-func (c *Cache) ResetFTWithMap(data map[string]map[string]interface{}, maxWords int, maxSizeBytes int, schema map[string]bool) error {
+func (c *Cache) ResetFTWithMap(data map[string]map[string]interface{}, maxWords int, maxSizeBytes uintptr, schema map[string]bool) error {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 	return c.resetFTWithMap(data, maxWords, maxSizeBytes, schema)
@@ -96,7 +96,7 @@ The FT cache is then reset and initialized with the data from the map.
 @Returns:
   - error: An error if the FullText cache is not initialized or if there is an error during the reset process
 */
-func (c *Cache) resetFTWithMap(data map[string]map[string]interface{}, maxWords int, maxSizeBytes int, schema map[string]bool) error {
+func (c *Cache) resetFTWithMap(data map[string]map[string]interface{}, maxWords int, maxSizeBytes uintptr, schema map[string]bool) error {
 	// If the FT cache is not initialized, return an error
 	if c.ft == nil {
 		return fmt.Errorf("full text is not initialized")
@@ -121,7 +121,7 @@ If an error occurs during the process, it will return an error.
 @Returns:
   - error: If the FullText cache has not been initialized, or if an error occurs during the reset and initialization process, it will return an error. Otherwise, it will return nil.
 */
-func (c *Cache) ResetFTWithJson(file string, maxWords int, maxSizeBytes int, schema map[string]bool) error {
+func (c *Cache) ResetFTWithJson(file string, maxWords int, maxSizeBytes uintptr, schema map[string]bool) error {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 	return c.resetFTWithJson(file, maxWords, maxSizeBytes, schema)
@@ -141,7 +141,7 @@ Args:
 Returns:
   - error: An error if the FT cache is not initialized, or an error occurs during initialization.
 */
-func (c *Cache) resetFTWithJson(file string, maxWords int, maxSizeBytes int, schema map[string]bool) error {
+func (c *Cache) resetFTWithJson(file string, maxWords int, maxSizeBytes uintptr, schema map[string]bool) error {
 	// If the FT cache is not initialized, return an error
 	if c.ft == nil {
 		return fmt.Errorf("full text is not initialized")
